requirements.txt
aiogram>=3.4.0
Pillow>=10.0.0
python-dotenv==1.2.2